+++ src/core/api/lookups/unidad-rol/api/unidad-rol.api-model.ts
@@ -1,0 +1,6 @@
+import { Lookup } from '#common/models';
+
+export interface UnidadRolList {
+  roles: Lookup[];
+  unidades: Lookup[];
+}
