+++ src/core/api/lookups/unidad-rol/api/unidad-rol.api.ts
@@ -1,0 +1,7 @@
+import axios from 'axios';
+import { UnidadRolList } from './unidad-rol.api-model';
+
+export const getUnidadRolList = async (): Promise<UnidadRolList> => {
+  const response = await axios.get<UnidadRolList>('/api/lookup/unidad-rol');
+  return response.data;
+};
