+++ src/core/api/lookups/unidad-rol/index.ts
@@ -1,0 +1,4 @@
+export * from './unidad-rol-vm';
+export * from './unidad-rol.mapper';
+export * from './unidad-rol.repository';
+export * from './use-unidad_rol-list.hook';
