+++ src/core/api/lookups/unidad-rol/unidad-rol-vm.ts
@@ -1,0 +1,11 @@
+import { Lookup } from '#common/models';
+
+export interface UnidadRolList {
+  roles: Lookup[];
+  unidades: Lookup[];
+}
+
+export const createEmptyUnidadRolList = (): UnidadRolList => ({
+  roles: [],
+  unidades: [],
+});
