+++ src/modules/users/create/api/create.api.ts
@@ -1,10 +1,5 @@
 import axios from 'axios';
-import { UnidadRolList, Usuario } from './create.api-model';
-
-export const getUnidadRolList = async (): Promise<UnidadRolList> => {
-  const response = await axios.get<UnidadRolList>('/api/lookup/unidad-rol');
-  return response.data;
-};
+import { Usuario } from './create.api-model';
 
 export const saveUser = async (user: Usuario): Promise<boolean> => {
   const response = await axios.post<boolean>('/api/user', user);
