+++ src/modules/users/create/create.pod.tsx
@@ -1,8 +1,9 @@
 import React from 'react';
 import { Spinner } from '#common/components';
 import { CreateUser } from './create.component';
-import { useSaveUserMutation, useUnidadRolList } from './create.query.hook';
+import { useSaveUserMutation } from './create.query.hook';
 import { Usuario } from './create.vm';
+import { useUnidadRolList } from '#core/api/lookups/unidad-rol';
 
 export const CreateUserPod: React.FC = () => {
   const { unidadRolList, isLoading } = useUnidadRolList();
